pyproject.toml
[project]
name = "mini-data-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.2",
    "dbt-duckdb>=1.10.0",
    "duckdb>=1.4.1",
    "pyyaml>=6.0",
]
